+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   Update Atomist package dependencies.
 -   Update TypeScript package dependencies.
+-   Migrate from teamIds to workspaceIds.
+-   **BREAKING** Removing the last rule from an ingress now returns an
+    ingress patch with a spec with a zero-length rules array.
 
 ## [0.8.0](https://github.com/atomist/k8-automation/compare/0.7.3...0.8.0) - 2018-05-18